+++ CHANGES.md
@@ -43,6 +43,7 @@
 * Compiler: directly write Wasm binary modules (#2000, #2003)
 * Compiler: rewrote inlining pass (#1935, #2018, #2027)
 * Compiler/wasm: optimize integer operations (#2032)
+* Compiler: Propagate arity between compilation units (#1594)
 
 ## Bug fixes
 * Compiler: fix stack overflow issues with double translation (#1869)